Fix #120119: Area lamp artifacts in Cycles with light trees
This is a regression in 4.1, caused by 36e603c430.

On a code side the issue was caused by a change in the mis_origin_n
which is a normal at a scatter point for light sampling.

Details about that specific change are a bit unclear. If not an
overseen mistake, it could have been part of idea that when no
bump map correction is used artists should have fuller control over
shading normal.

In any case, the change to this normal lead to some discrepancies
with shading/sampling/NEE code which uses sd->N and has no access
to closure.

In order to bring code back to consistent state this change reverts
the change to mis_origin_n. However it also does some extra things
to help with the original file used during development the bump map
correction option, as without doing anything else the file renders
with firefly-looking artifacts.

Those artifacts are due to normal going into the surface. While the
original thought was to let artists to have full control over the
normals and deal with such cases from shader, it seems that it
happens more often, could be hard to understand by artists what's
going on, and is also not compatible with features like light
trees.

Hence this change also does partial bump map correction when the
option is disabled: namely it forces evaluation to be 0 when normal
goes below the surface. This minimizes an impact on the files which
are currently relying on the dump map corrections being disabled.