Fix #120212: GPv3: Hide masks flag wrongly set
This was caused by ab84f9eaf062f30b88acdb46415c35b42e4b6e8c.

The issue is that setting the flag in the `TreeNode`
constructor means that the root group will always
hide masks when its created. This effectively disables
masks in the entire object.

Instead, leave the default flag to be `0` (masks enabled)
and explicitly set the `HIDE_MASKS` flag when
creating a new layer/layer group.