Use the gather function to determine group interface.
The states of internal nodes include superpositions of single value and
field to indicate that a socket _can_ be either, but the overall group
interface should only include the concrete connected types. A single
value socket connected to a group output creates a single-or-field
socket, while the group interface should consider it a single value.
The `find_group_output_dependencies` has to be used to determine the
interface types for the tree as a whole, rather than the state of
internal sockets.