Fix #117338: Texture paint sampling broken with modifiers
e64b3c821252fc8707b4 worked to complete a transition started during 2.8
development to avoid dynamically re-evaluating objects outside of the
dependency graph's control. However, that behavior was used to make
sure the original index mapping was available for the texture paint
sample operator.

Conceptually, sampling a texture on an arbitrary evaluated mesh should
not require original indices at all though. All we need to know is the
UV map value under the mouse. This commit changes from using GPU index
textures to a BVH tree raycast. This significantly simplifies the code
and makes it work with GPU subdivision too. Though it may be slower to
build the BVH tree, that should be okay because it's cached and only
needs to be built once, and that's something we're looking to optimize
in the future anyway.

Removing the reliance on original indices also means we could paint
textures on completely procedurally generated meshes, which may be
an interesting feature in the future.