Fix #120239: Snap in edit mode ignoring some elements
Caused by 1c77779160 when we create a mesh representing the edited
mesh.

Caused by 1c77779160 when we create a mesh representing the edited
mesh.

Mesh Snap to Vertex works in the following order:
- Snap to vertices of visible triangles
- Snap to vertices of Loose Edges
- Snap to Loose Vertices

The problem happens because, in editing mode, we ignore faces whose
vertices are being transformed. We mark this face as being hidden in
the snap. So we lost some vertices in triangles.

The solution then is to consider the edges and vertices of hidden faces
as loose elements, because, although connected to faces, they are still
visible to snap.

Therefore, two new types of BVHTree were created:
- BVHTREE_FROM_LOOSEVERTS_NO_HIDDEN
- BVHTREE_FROM_LOOSEEDGES_NO_HIDDEN

The disadvantage of this solution is that we can no longer rely on
cached `mesh->loose_verts()` and `mesh->loose_edges()` which can be bad
for performance.