Fix #120287: GPv3: Crash when converting file with annotations
The `greasepencil::convert::legacy_main` function did not account for annotations.
They are `bGPData` IDs but shouldn't be converted to GPv3.

The fix does this by skipping over `bGPData` IDs that have the
`GP_DATA_ANNOTATIONS` flag set.
Note that this also covers the case where an ID is used
both as annotation and object data. In this case, the ID
remapping is skipped, but that is ok, because the
legacy data still has at least one user (annotations).