Fix: Prevent from Liquid leaking through Effector with standard settings.
When the liquid is created with some collision/effector with default settings,
it is often happen that the liquid is leaking through the Solid object/Effector.

It is counter intuitive, and it is causing frutstration for new users.

To resolve that issue, the CFL numer could be decreased (maximal value of
velocity on which additional step will be added).
After decreasing it from 4.0 to 2.0, the leaking issue is fixed.
Additionally there is no noticable change in performance.
I have tested it on different resolutions (32, 64, 100, 128).