Fix #118694: snap cursor sometimes doesn't update when dragging and dropping assets
The problem happens because, on some OSs, `wm->winactive->eventstate`
does not update when the mouse moves. So the snap cursor doesn't update
because apparently the mouse didn't move.

The solution in this commit is to use the `eventstate` of the window in
the context (`CTX_wm_window(C)->eventstate`).

This matches other code snippets where `eventstate` is used.

---
The behavior on each OS apparently happens like this:
**Windows:**
- `wm->winactive` is updated if the cursor is over the window for a little less than 1 second
- Events are sent to the window whose cursor is over.

**MacOS:**
- `wm->winactive` is the last one selected
- Events are sent to the active window and to the one whose cursor is over.

**Linux:**
- `wm->winactive` is the last one selected
- Events are sent to the window whose cursor is over.

Therefore, we cannot trust the eventstate of `wm->winactive`.