Fix #120046: Broken render on AMD Metal GPUs
This is a regression since fdc2962bebb

The size of state can not be different between CPU and GPU.

This change replaces compile-time condition with a kernel feature
check, which solves the render regression on AMD Metal. It also
minimizes the state size on other GPUs when Light Tree is disabled.