WM: add preference for wayland to force the scroll direction
Even though the wayland protocol has the ability to set the scroll
direction, it seems none of the mainstream wayland compositors support
this. GNOME 46 and KDE 6, WLROOTS appliers to support this although
the latest version of SWAYWM doesn't work in my tests.

This workaround can be removed once the underlying API's are better
supported.