Fix: Move CFL Number parameter for Fluid into Adaptive Time Steps
The CFL Number is only used, when Adaptive Time Steps is enabled.
Currently it is not part of Adaptive Time Steps, and changing it
doesn't have any effect, if Adaptive Time Steps is disabled.

With this commit, the CFL parameter was moved into Adaptive Time Steps.

More information about CFL number is available at:
https://github.com/thunil/mantaflow/blob/494445143507296d16e467160c1889be657f8f74/source/fluidsolver.h#L62