Subdiv: Remove topology refiner C-API wrapper
Remove the indirection previously used for the topology refiner
to separate C and C++ code. Instead retrieve the base level in
calling code and call opensubdiv API functions directly. This
avoids copying arrays of mesh indices and should reduce
function call overhead since index retrieval is now inlined.
It also lets us remove a lot of boilerplate shim code.

Similar to the previous commit, the downside is increased need
for WITH_OPENSUBDIV defines in various parts of blenkernel,
but I think that is required to avoid the previous indirection.