UI: Enable Mesh displaying for Liquid
Problem:
With previous implementation Mesh generation is disabled for Liquid.

It is causing confusing for new users, as default Liquid Particle System,
is visible only with X-Ray enabled with Wireframe and Solid view.

Particles are not visible in Material Preview and Rendered view.
It is cauing frustration from new users, and feeling that Blender
is hard to use.

Solution:
Generate and display Mesh for Liquid by default.

For new user, it would be much easier to setup the liquid simulation

Advanced users should not have problems with disabling Mesh generation,
and tune other parameters for their needs.