Fix #120918: EEVEE-Next world volume on iGPU HD530
When using world shaders with volume, older hardware could create
errors when the world background wasn't behind a surface. The reason
is that the world volume shader had a color attachment attached, but
none were defined by the shader interface. This lead to undefined
behavior.

This PR fixes this by attaching the occupancy frame buffer. This
framebuffer only contains a depth attachment.