Geometry Nodes: Character at Index node
Add a new string node that gives the character at a specific index
of a string. While the strings are processed internally as UTF8, the
indexed unicode character is returned, both as a string, and as the
unicode character code integer.

The initial use case is as a general building block for calculating
information like the word index. When used in combination with
the string to curves node and the accumulate field node, this can
be helpful for various motion graphics effects.

When the input index is invalid, currently -1 and an empty string
are returned.