GPU: Fix out of bound write when logging system paths
Blender expects that only the filename is provided which is true
for internal shader sources.

Metal shaders can error and return a full path to a system shader.
This doesn't fit inside the reserved memory that Blender reserved
for logging filename and line number.

This out of bound write can be easilly triggered when using `min`
where the parameters aren't of the same kind.
`uint min(uint, int)` for example.

This PR reseves more space to store the filename.