Refactor: Blend file reading asserts for broken pointers
Previously these were silently ignored. In general they should not happen,
but due to versioning and bugs some blend files do have broken pointers.

The known cases are explicitly marked as such now, and the assert is skipped.
Those cases must correctly handle the pointer being null.

Future Blender versions are likely to write more invalid pointers, which
would then trigger these asserts in older versions. I guess as long as
it is only an assert this is ok, but could be annoying in some cases.
Though with these asserts they are more likely to be caught and fixed.