IMB: Cleanup timecode options
Remove options that are duplicate, rename remaining options, so these
represent to how timecodes are actually used.

Timecodes in Blender have pretty much nothing in common with more widely
known term "timecode". This confused users (and developers).
There were 5 options of which 3 were exactly same. This commit leaves
user with 3 options:
 - Play movie normally - "Use Movie Timestamps" in menu
 - Ensure, that no frame is duplicated or skipped - "No Gaps Playback"
 - Do not use timecodes - "None"

Description of these options now describes, what these do, and
implicitly, when they should be used.

Perhaps the intention was, that if your device uses record run
or free run timecode, you could pick this option and it would just
work. But this fact and the description of these options is just
misleading.

----

Perhaps this could be more verbose in description, For example "Use Movie Timestamps" should be used if there are issues with seeking.

The DNA naming could be more precise as well - IMB_TC_INVERSE_MAPPING means, that timestamp -> frame table is generated by info from ffmpeg, we than use it as frame -> timestamp It could be called IMB_TC_USE_FRAME_PTS_MAP, but I think, that in some streams, instead of PTS, a binary offset is used. So inverse mapping is more general, but not telling much to person not knowing anything about this system.

The whole term Timecode could be renamed as well, at least in UI, so this is even less confusing