WIP: VSE: Draw outline on active unselected strips
In the VSE, strips can be active but not selected. This means that
their content shows up in the sidebar properties, but they do not react
to transform shortcuts or other operations.

Draw a subtle 1px outline on active unselected strips, using the color of
the active outline (but dimmed slightly).