EEVEE-Next: Improve shadow bias
Compute shadow bias in the normal direction to avoid
both shadow leaking at certain angles or shadow
acnee.

This bias is computed automatically based on the minimum
bias amount required to remove all errors.

We do the bias in world space instead of shadow space
for speed and simplicity. This requires us to bias using
the upper bound of biases for the same location in space
(using biggest texel world radius instead of UVZ bias).
This isn't much of an issue since the bias is still less
than 2 texel. The bias is still modulated by facing
ratio to the light so surfaces facing the light have no
biases.

We always bias toward the normal direction instead of the
light direction. This is alike Cycles geometric offset for
the shadow terminator fix. This is better since it does'nt
modify the shading at all.

This changes the PCF filtering to only filter after the
the "safe" zone where we cannot have any shadow acnee.
This introduces bias in this regard since we move the
shadow map center away from the surface.

This removes the geometric offset inside the shadow surface
shader as we already account for that at sampling time.